requirements.txt
pydantic==2.10.6
pillow==10.4.0
moviepy==2.1.2
numpy==2.0.2
regex==2024.11.6
requests==2.32.3
websockets==14.2
flask==3.1.0
gevent==24.11.1
matplotlib==3.9.4